+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## UNRELEASED
 
-  * Added :max_cookies and :max_cookies_per_domain options
-  * Added ReqCookieManager plugin
+  * Added `:max_cookies` and `:max_cookies_per_domain` options
+  * Added `HttpCookie.ReqPlugin`
   * Fixed updating last access time
 
 ## v0.6.0 (2024-01-23)
@@ -13,7 +13,7 @@
 
 ## v0.5.1 (2024-01-20)
 
-  * Changed nimble_parsec to dev-only dependency
+  * Changed `nimble_parsec` to dev-only dependency
   * Added more typespecs
 
 ## v0.5.0 (2024-01-16)